+++ src/main/java/de/maxhenkel/plane/entity/EntityCargoPlane.java
@@ -110,7 +110,7 @@ public class EntityCargoPlane extends EntityPlaneBase {
 
     @Override
     public int getFuelCapacity() {
-        return 16000;
+        return Main.SERVER_CONFIG.cargoPlaneFuelCapacity.get();
     }
 
     @Override