+++ changelog.md
@@ -1,1 +1,2 @@
 - Fixed plane GUI showing when having the GUI hidden
+- Made plane fuel capacity configurable